# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Piece-Meal App
# during dev, should point to your local machine
# host: piecemeal-api.herokuapp.com
host: localhost:10010
...
basePath: /api/v1
...
schemes:
  # tip: remove http to make production-grade
  # - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /ingredients:
    x-swagger-router-controller: ingredients
    get:
      description: get the ingredients list
      # define the type of response for Success "200" and Error
      operationId: getIngredientsList
      # parameters:
      #   - name: token
      #     in: header
      #     description: Users authentication token
      #     required: true
      #     type: string

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIngredientsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    post:
      description: add a new ingredient to the list
      # movie info to be stored
      operationId: addIngredient
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: ingredient
          description: Ingredient
          in: body
          required: true
          schema:
            $ref: "#/definitions/Ingredient"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIngredientResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /ingredients/{id}:
    x-swagger-router-controller: ingredients
    get:
      description: get an ingredient
      # define the type of response for Success "200" and Error
      operationId: getIngredient
      parameters:
        # - name: token
        #   in: header
        #   description: Users authentication token
        #   required: true
        #   type: string

        - name: id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIngredientResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    put:
      description: update an ingredient
      # define the parameters
      operationId: updateIngredient
      parameters:
        - name: id
          type: integer
          in: path
          required: true
        - name: ingredient
          description: Ingredient
          in: body
          required: true
          schema:
            $ref: "#/definitions/PatchIngredient"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIngredientResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: delete an ingredient
      # define the parameters
      operationId: deleteIngredient
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: id
          description: Ingredient id
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIngredientResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /ingredients/{id}/alternatives:
    x-swagger-router-controller: ingredients
    # get:
    #   description: get alternatives for an ingredient
    #   # define the type of response for Success "200" and Error
    #   operationId: getIngredientAlternatives
    #   parameters:
    #     - name: id
    #       type: integer
    #       in: path
    #       required: true
    #   responses:
    #     "200":
    #       description: Success
    #       schema:
    #         $ref: "#/definitions/GetIngredientAlternativesResponse"
    #     default:
    #       description: Error
    #       schema:
    #         $ref: "#/definitions/ErrorResponse"

    post:
      description: add alternatives for an ingredient
      # define the type of response for Success "200" and Error
      operationId: addIngredientAlternatives
      parameters:
        - name: token
          in: header
          description: Users authentication token
          required: true
          type: string

        - name: id
          type: integer
          in: path
          required: true
        - name: alternatives
          in: body
          description: alternaive ingredients
          # type: array
          # items:
          #   type: integer
          schema:
            type: object
            required:
              - ingredients
            properties:
              ingredients:
                description: array of alt ingredient ids
                type: array
                items:
                  type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIngredientResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /search/ingredients:
    x-swagger-router-controller: ingredients
    get:
      description: get results from ingredient search
      operationId: searchIngredients
      parameters:
        # - name: token
        #   in: header
        #   description: Users authentication token
        #   required: true
        #   type: string

        - name: text
          type: string
          in: query
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetIngredientsListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  PatchIngredient:
    type: object
    properties:
      id:
        description: Ingredient id
        type: integer
      name:
        description: Ingredient name
        type: string
      description:
        description: Ingredient description
        type: string
      tags:
        description: Ingredient tags
        type: array
        items:
          type: string
      alternatives:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
      image_url:
        type: string
        description: link to image of ingredient
  Ingredient:
    type: object
    properties:
      id:
        type: number
        description: ingredient id
      name:
        type: string
        description: ingredient object name
      description:
        type: string
        description: description of ingredient
      tags:
        type: array
        items:
          type: string
      alternatives:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
      image_url:
        type: string
        description: link to image of ingredient
    required:
      - name
...
  GetIngredientsListResponse:
    required:
      - ingredients
    properties:
      # The array of ingredients
      ingredients:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            tags:
              type: array
              items:
                type: string
            alternatives:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
            image_url:
              type: string
              description: link to image of ingredient
...
  GetIngredientResponse:
    required:
      - id
      - name
      # - alternatives
      # - tags
    properties:
      id:
        type: integer
      name:
        type: string
      active:
        type: boolean
      tags:
        type: array
        items:
          type: string
      alternatives:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
      calories:
        type: integer
      image_url:
        type: string
        description: link to image of ingredient

  # GetIngredientAlternativesResponse:
  #   required:
  #     - id
  #     - name
  #     - alternatives
  #   properties:
  #     id:
  #       type: number
  #     name:
  #       type: string
  #     alternatives:
  #       type: array
  #       items:
  #         $ref: "#/definitions/Ingredient"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...